view/requirements.txt
gradio
torch
torchvision
Pillow
concrete-ml==1.6.1